feedr/requirements.txt
Django==1.6.5
South==1.0
django-debug-toolbar==1.2.1
facebook-sdk==0.4.0
oauthlib==0.6.3
psycopg2==2.5.3
python-openid==2.2.5
python-social-auth==0.1.26
requests==2.3.0
requests-oauthlib==0.4.1
six==1.7.3
sqlparse==0.1.11
wsgiref==0.1.2
